Removed db access for easy deployment

+++ src/main/java/com/example/springboot/service/MealService.java
@@ -8,6 +8,8 @@ import com.mashape.unirest.http.JsonNode;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
 import org.json.JSONObject;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,6 +20,7 @@ import java.util.List;
 public class MealService {
     private MealDao mealDao;
     private Double MIN_SUMMER_TEMP = 20.0;
+    private static final Logger LOGGER = LoggerFactory.getLogger(MealService.class);
 
     @Autowired
     public MealService(MealDao mealDao) {
@@ -33,6 +36,13 @@ public class MealService {
     }
 
     public void deleteMeal(long id) {
+        List<Meal> mealsById = mealDao.findById(id);
+
+        if (mealsById.isEmpty()) {
+            LOGGER.warn("Attempt to delete non-existing meal with id: " + id);
+            return;
+        }
+
         mealDao.deleteById(id);
     }
